#5719 pdate dp-api-clients-go method to reflect updates to GET areas endpoint and point to endpoint in dp-population-types-api
