+++ dimension/client.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/ONSdigital/dp-api-clients-go/v2/clientlog"
 	dperrors "github.com/ONSdigital/dp-api-clients-go/v2/errors"
 	"github.com/ONSdigital/dp-api-clients-go/v2/headers"
 	"github.com/ONSdigital/dp-api-clients-go/v2/health"
@@ -112,62 +111,6 @@ func checkGetResponse(resp *http.Response) error {
 	}
 
 	return nil
-}
-
-func (c *Client) GetAreas(ctx context.Context, input GetAreasInput) (GetAreasResponse, error) {
-	logData := log.Data{
-		"method":       http.MethodGet,
-		"dataset_id":   input.DatasetID,
-		"area_type_id": input.AreaTypeID,
-		"text":         input.Text,
-	}
-
-	urlPath := "areas"
-	urlValues := url.Values{"dataset": []string{input.DatasetID}, "area-type": []string{input.AreaTypeID}}
-	if input.Text != "" {
-		urlValues.Add("text", input.Text)
-	}
-
-	req, err := c.createGetRequest(ctx, input.UserAuthToken, input.ServiceAuthToken, urlPath, urlValues)
-	if err != nil {
-		return GetAreasResponse{}, dperrors.New(
-			err,
-			dperrors.StatusCode(err),
-			logData,
-		)
-	}
-
-	clientlog.Do(ctx, "getting areas", service, req.URL.String(), logData)
-
-	resp, err := c.hcCli.Client.Do(ctx, req)
-	if err != nil {
-		return GetAreasResponse{}, dperrors.New(
-			errors.Wrap(err, "failed to get response from Dimensions API"),
-			http.StatusInternalServerError,
-			logData,
-		)
-	}
-
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			log.Error(ctx, "error closing http response body", err)
-		}
-	}()
-
-	if err := checkGetResponse(resp); err != nil {
-		return GetAreasResponse{}, err
-	}
-
-	var areas GetAreasResponse
-	if err := json.NewDecoder(resp.Body).Decode(&areas); err != nil {
-		return GetAreasResponse{}, dperrors.New(
-			errors.Wrap(err, "unable to deserialize areas response"),
-			http.StatusInternalServerError,
-			logData,
-		)
-	}
-
-	return areas, nil
 }
 
 // newRequest creates a new http.Request with auth headers
