+++ dimension/client_test.go
@@ -1,12 +1,9 @@
 package dimension
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"testing"
 
@@ -33,212 +30,6 @@ func TestNewClient(t *testing.T) {
 
 		Convey("the constructor should return an error", func() {
 			So(err, ShouldBeError)
-		})
-	})
-}
-
-func TestGetAreas(t *testing.T) {
-	Convey("Given a valid request", t, func() {
-		stubClient := newStubClient(&http.Response{Body: ioutil.NopCloser(bytes.NewReader(nil))}, nil)
-		client, err := NewWithHealthClient(health.NewClientWithClienter("", "http://test.test:2000/v1", stubClient))
-		So(err, ShouldBeNil)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "testText",
-		}
-		_, _ = client.GetAreas(context.Background(), input)
-
-		Convey("it should call the areas endpoint, serializing the dataset, area type and text query params", func() {
-			calls := stubClient.DoCalls()
-			So(calls, ShouldNotBeEmpty)
-			So(calls[0].Req.URL.String(), ShouldEqual, "http://test.test:2000/v1/areas?area-type=testAreaType&dataset=testDataSet&text=testText")
-		})
-	})
-
-	Convey("Given a valid request with an empty text param", t, func() {
-		stubClient := newStubClient(&http.Response{Body: ioutil.NopCloser(bytes.NewReader(nil))}, nil)
-		client, err := NewWithHealthClient(health.NewClientWithClienter("", "http://test.test:2000/v1", stubClient))
-		So(err, ShouldBeNil)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-
-		_, _ = client.GetAreas(context.Background(), input)
-
-		Convey("it should call the areas endpoint, omitting the text query param", func() {
-			calls := stubClient.DoCalls()
-			So(calls, ShouldNotBeEmpty)
-			So(calls[0].Req.URL.String(), ShouldEqual, "http://test.test:2000/v1/areas?area-type=testAreaType&dataset=testDataSet")
-		})
-	})
-
-	Convey("Given authentication tokens", t, func() {
-		const userAuthToken = "user"
-		const serviceAuthToken = "service"
-
-		stubClient := newStubClient(&http.Response{Body: ioutil.NopCloser(bytes.NewReader(nil))}, nil)
-		client := newHealthClient(stubClient)
-
-		input := GetAreasInput{
-			UserAuthToken:    userAuthToken,
-			ServiceAuthToken: serviceAuthToken,
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-
-		_, _ = client.GetAreas(context.Background(), input)
-
-		Convey("it should set the auth headers on the request", func() {
-			calls := stubClient.DoCalls()
-			So(calls, ShouldNotBeEmpty)
-
-			So(calls[0].Req, shouldHaveAuthHeaders, userAuthToken, serviceAuthToken)
-		})
-	})
-
-	Convey("Given a valid areas response payload", t, func() {
-		areas := GetAreasResponse{
-			Areas: []Area{{ID: "test", Label: "Test", AreaType: "city"}},
-		}
-
-		resp, err := json.Marshal(areas)
-		So(err, ShouldBeNil)
-
-		stubClient := newStubClient(&http.Response{
-			StatusCode: http.StatusOK,
-			Body:       ioutil.NopCloser(bytes.NewReader(resp)),
-		}, nil)
-		client := newHealthClient(stubClient)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-		types, err := client.GetAreas(context.Background(), input)
-
-		Convey("it should return a list of areas", func() {
-			So(err, ShouldBeNil)
-			So(types, ShouldResemble, areas)
-		})
-	})
-
-	Convey("Given the dimensions API returns an error", t, func() {
-		stubClient := newStubClient(nil, errors.New("oh no"))
-
-		client := newHealthClient(stubClient)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-		_, err := client.GetAreas(context.Background(), input)
-
-		Convey("it should return an internal error", func() {
-			So(err, shouldBeDPError, http.StatusInternalServerError)
-		})
-	})
-
-	Convey("Given the dimensions API returns a status code of 404", t, func() {
-		stubClient := newStubClient(&http.Response{
-			StatusCode: http.StatusNotFound,
-			Body:       ioutil.NopCloser(bytes.NewReader([]byte(`{ "errors": ["not found"] }`))),
-		}, nil)
-
-		client := newHealthClient(stubClient)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-		_, err := client.GetAreas(context.Background(), input)
-
-		Convey("the error chain should contain the original Errors type", func() {
-			So(err, shouldBeDPError, http.StatusInternalServerError)
-
-			var respErr ErrorResp
-			ok := errors.As(err, &respErr)
-			So(ok, ShouldBeTrue)
-			So(respErr, ShouldResemble, ErrorResp{Errors: []string{"not found"}})
-		})
-	})
-
-	Convey("Given the dimensions API returns a status code other than 200/400", t, func() {
-		stubClient := newStubClient(&http.Response{
-			StatusCode: http.StatusBadRequest,
-			Body:       ioutil.NopCloser(bytes.NewReader([]byte(`{ "areas": [] }`))),
-		}, nil)
-
-		client := newHealthClient(stubClient)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-		_, err := client.GetAreas(context.Background(), input)
-
-		Convey("it should return an internal error", func() {
-			So(err, shouldBeDPError, http.StatusInternalServerError)
-		})
-	})
-
-	Convey("Given the response cannot be deserialized", t, func() {
-		stubClient := newStubClient(&http.Response{
-			StatusCode: http.StatusOK,
-			Body:       ioutil.NopCloser(bytes.NewReader([]byte(`{ "areas" `))),
-		}, nil)
-
-		client := newHealthClient(stubClient)
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-		_, err := client.GetAreas(context.Background(), input)
-
-		Convey("it should return an internal error", func() {
-			So(err, shouldBeDPError, http.StatusInternalServerError)
-		})
-	})
-
-	Convey("Given the request cannot be created", t, func() {
-		client := newHealthClient(newStubClient(nil, nil))
-
-		input := GetAreasInput{
-			UserAuthToken:    "",
-			ServiceAuthToken: "",
-			DatasetID:        "testDataSet",
-			AreaTypeID:       "testAreaType",
-			Text:             "",
-		}
-		_, err := client.GetAreas(nil, input)
-
-		Convey("it should return a client error", func() {
-			So(err, shouldBeDPError, http.StatusBadRequest)
 		})
 	})
 }
